Align implementation with hdpi 774
