+++ src/main/java/uk/gov/hmcts/reform/enforcement/notify/config/NotificationErrorHandler.java
@@ -15,6 +15,16 @@ import java.util.function.Consumer;
 @Slf4j
 public class NotificationErrorHandler {
 
+    /**
+     * Handles exceptions occurring during the process of sending an email notification.
+     * Based on the HTTP status code of the exception, this method updates the notification's
+     * status and may throw further exceptions if necessary.
+     *
+     * @param exception the {@code NotificationClientException} encountered while sending the email
+     * @param caseNotification the {@code CaseNotification} object containing details about the notification
+     * @param referenceId a unique reference ID corresponding to the notification attempt
+     * @param statusUpdater a {@code Consumer} function to update the status of the notification
+     */
     public void handleSendEmailException(NotificationClientException exception,
                                             CaseNotification caseNotification,
                                             String referenceId,
@@ -54,6 +64,14 @@ public class NotificationErrorHandler {
         }
     }
 
+    /**
+     * Handles exceptions that occur during the fetch notification process by logging the error details
+     * and rethrowing a custom {@code NotificationException}.
+     *
+     * @param exception       The exception thrown during the fetch operation, containing details such as HTTP status
+     *                         code.
+     * @param notificationId  The unique identifier of the notification that failed to fetch.
+     */
     public void handleFetchException(NotificationClientException exception, String notificationId) {
         int httpStatusCode = exception.getHttpResult();
 
@@ -67,6 +85,11 @@ public class NotificationErrorHandler {
         throw new NotificationException("Failed to fetch notification, please try again.", exception);
     }
 
+    /**
+     * Represents an update to the status of a case notification.
+     * This class encapsulates information about the notification,
+     * the updated status, and an optional provider notification ID.
+     */
     public record NotificationStatusUpdate(CaseNotification notification, NotificationStatus status,
                                             UUID providerNotificationId) {
     }
