+++ src/main/java/uk/gov/hmcts/reform/enforcement/notify/task/SendEmailTaskComponent.java
@@ -147,7 +147,7 @@ public class SendEmailTaskComponent {
             });
     }
 
-    void updateNotificationFromStatusUpdate(NotificationStatusUpdate statusUpdate) {
+    private void updateNotificationFromStatusUpdate(NotificationStatusUpdate statusUpdate) {
         notificationService.updateNotificationStatus(
             statusUpdate.notification().getNotificationId(),
             statusUpdate.status().toString()
