+++ src/test/java/uk/gov/hmcts/reform/enforcement/testingsupport/endpoint/TestingSupportControllerTest.java
@@ -12,16 +12,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.reform.enforcement.notify.exception.NotificationException;
-import uk.gov.hmcts.reform.enforcement.notify.model.EmailNotificationRequest;
-import uk.gov.hmcts.reform.enforcement.notify.model.EmailNotificationResponse;
-import uk.gov.hmcts.reform.enforcement.notify.model.NotificationStatus;
-import uk.gov.hmcts.reform.enforcement.notify.service.NotificationService;
 
 import java.time.Instant;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -32,9 +24,6 @@ import static org.mockito.Mockito.when;
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("unchecked")
 class TestingSupportControllerTest {
-
-    @Mock
-    private NotificationService notificationService;
 
     @Mock
     private SchedulerClient schedulerClient;
@@ -48,8 +37,6 @@ class TestingSupportControllerTest {
     @InjectMocks
     private TestingSupportController testingSupportController;
 
-    private EmailNotificationRequest emailRequest;
-    private EmailNotificationResponse emailResponse;
     private String authorization;
     private String serviceAuthorization;
 
@@ -57,79 +44,12 @@ class TestingSupportControllerTest {
     void setUp() {
         authorization = "Bearer token123";
         serviceAuthorization = "ServiceAuth token456";
-
-        Map<String, Object> personalisation = new HashMap<>();
-        personalisation.put("name", "John Doe");
-        personalisation.put("reference", "REF123");
-
-        emailRequest = new EmailNotificationRequest();
-        emailRequest.setEmailAddress("test@example.com");
-        emailRequest.setTemplateId("template-123");
-        emailRequest.setPersonalisation(personalisation);
-        emailRequest.setReference("notification-ref");
-        emailRequest.setEmailReplyToId("reply-to-123");
-
-        emailResponse = new EmailNotificationResponse();
-        emailResponse.setTaskId("task-123");
-        emailResponse.setStatus(NotificationStatus.SCHEDULED.toString());
-        emailResponse.setNotificationId(UUID.randomUUID());
-    }
-
-    @Test
-    void sendEmail_ShouldReturnAccepted_WhenEmailScheduledSuccessfully() {
-        when(notificationService.scheduleEmailNotification(emailRequest)).thenReturn(emailResponse);
-
-        ResponseEntity<EmailNotificationResponse> response = testingSupportController.sendEmail(
-            authorization, serviceAuthorization, emailRequest);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.ACCEPTED);
-        assertThat(response.getBody()).isEqualTo(emailResponse);
-        assertThat(response.getBody().getTaskId()).isEqualTo("task-123");
-        assertThat(response.getBody().getStatus()).isEqualTo(NotificationStatus.SCHEDULED.toString());
-
-        verify(notificationService).scheduleEmailNotification(emailRequest);
-    }
-
-    @Test
-    void sendEmail_ShouldReturnInternalServerError_WhenNotificationServiceThrowsException() {
-        when(notificationService.scheduleEmailNotification(emailRequest))
-            .thenThrow(new NotificationException("Failed to schedule notification",
-                                                    new RuntimeException("Database error")));
-
-        ResponseEntity<EmailNotificationResponse> response = testingSupportController.sendEmail(
-            authorization, serviceAuthorization, emailRequest);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR);
-        assertThat(response.getBody()).isNull();
-
-        verify(notificationService).scheduleEmailNotification(emailRequest);
-    }
-
-    @Test
-    void sendEmail_ShouldReturnInternalServerError_WhenRuntimeExceptionOccurs() {
-        when(notificationService.scheduleEmailNotification(emailRequest))
-            .thenThrow(new RuntimeException("Unexpected error"));
-
-        ResponseEntity<EmailNotificationResponse> response = testingSupportController.sendEmail(
-            authorization, serviceAuthorization, emailRequest);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR);
-        assertThat(response.getBody()).isNull();
-
-        verify(notificationService).scheduleEmailNotification(emailRequest);
-    }
-
-    @Test
-    void sendEmail_ShouldUseDefaultAuthorization_WhenHeaderNotProvided() {
-        when(notificationService.scheduleEmailNotification(emailRequest)).thenReturn(emailResponse);
-
-        ResponseEntity<EmailNotificationResponse> response = testingSupportController.sendEmail(
-            "DummyId", serviceAuthorization, emailRequest);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.ACCEPTED);
-        assertThat(response.getBody()).isEqualTo(emailResponse);
-
-        verify(notificationService).scheduleEmailNotification(emailRequest);
+        
+        // Need to fix the constructor for InjectMocks to work properly with the new signature
+        testingSupportController = new TestingSupportController(
+            schedulerClient,
+            helloWorldTask
+        );
     }
 
     @Test
@@ -271,7 +191,7 @@ class TestingSupportControllerTest {
     @Test
     void constructor_ShouldInitializeFields() {
         TestingSupportController controller = new TestingSupportController(
-            notificationService, schedulerClient, helloWorldTask);
+            schedulerClient, helloWorldTask);
 
         assertThat(controller).isNotNull();
     }
